llmcustomersupportapp/requirements.txt
flask
google.generativeai
python-dotenv
wtforms
IPython
flask_wtf
wtforms.validators
chromadb